Refactoring replaceOrAddProtector to keyContainer
